+++ app/Livewire/Product/DeleteProduct.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Product;
 
 use App\Models\Product;
+use App\Models\Request;
 use Livewire\Attributes\On;
 use Livewire\Component;
 use Mary\Traits\Toast;
@@ -30,6 +31,13 @@ class DeleteProduct extends Component
 
     public function delete()
     {
+        if (Request::where('product_id', $this->product)->exists()) {
+            $this->error('Products can not be deleted if they are used in a request');
+
+            $this->deleteProductModal = false;
+            return;
+        }
+
         Product::destroy($this->product);
 
         $this->deleteProductModal = false;
